+++ resources/views/components/menu.blade.php
@@ -23,22 +23,22 @@
                                      <ul class="nav navbar-nav">
                                          <li wire:click="changeView('main-page')">
                                              <a class="show-1 templatemo_home" href="/">
-                                                 <span class="fa fa-home"></span>Главная</a>
+                                                 <span class="fa fa-home"></span>{!! __('Главная') !!}</a>
                                          </li>
                                          @if (Auth::check())
                                              <li wire:click="changeView('subjects')">
                                                  <a class="show-1 templatemo_page2" href="/Subjects">
-                                                     <span class="fa fa-picture-o"></span>Дисциплины</a>
+                                                     <span class="fa fa-picture-o"></span>{!! __('Дисциплины') !!}</a>
                                              </li>
                                              <li>
                                                  <a href="logout">
-                                                     <span class="fa fa-users"> </span>Выйти
+                                                     <span class="fa fa-users"> </span>{!! __('Выйти') !!}
                                                  </a>
                                              </li>
                                          @else
                                              <li wire:click="changeView('login')">
                                                  <a class="show-1 templatemo_page3" href="/Login">
-                                                     <span class="fa fa-users"></span>Войти</a>
+                                                     <span class="fa fa-users"></span>{!! __('Войти') !!}</a>
                                              </li>
                                          @endif